Add signal system and signals
For now signals are immediately emitted and therefore we can just create
signals on the stack (they are immediately used and so their lifetime is
guaranteed even though we take a pointer parameter in EmitSignal)

In future we could queue signals for later and process and deque them in an
update loop. In this case, we would have to dynamically allocate signals. The
performance hit would probably be fine. A linear allocator would be the
solution if all the allocations started to take up time

See Elias Daler's Bikeshed Renderer for where this all comes from